requirements.txt
# Core Data Science & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pypdf>=3.17.0

# NLP & Text Processing
nltk>=3.8.0
spacy>=3.7.0
langdetect>=1.0.9

# Embeddings & Transformers
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
tokenizers>=0.15.0

# Retrieval Systems
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.22
qdrant-client>=1.7.0

# LLM Integration
openai>=1.10.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Evaluation Metrics
rouge-score>=0.1.2
bert-score>=0.3.13
ragas>=0.1.0
deepeval>=0.20.0

# Web Interface & API
streamlit>=1.29.0
fastapi>=0.109.0
uvicorn>=0.25.0
gradio>=4.16.0

# Data Handling & Storage
jsonlines>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

# Visualization & Analysis
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
tqdm>=4.66.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
urllib3>=2.1.0

# Development & Testing
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0

# Optional: GPU Support (uncomment if using GPU)
# faiss-gpu>=1.7.4
# torch>=2.0.0+cu118
